+++ index.js
@@ -1,10 +1,10 @@
 const path = require('path');
 const core = require('@actions/core');
-const aws = require('aws-sdk');
+const { CodeDeploy, waitUntilDeploymentSuccessful } = require('@aws-sdk/client-codedeploy');
+const { ECS, waitUntilServicesStable, waitUntilTasksStopped } = require('@aws-sdk/client-ecs');
 const yaml = require('yaml');
 const fs = require('fs');
 const crypto = require('crypto');
-
 const MAX_WAIT_MINUTES = 360;  // 6 hours
 const WAIT_DEFAULT_DELAY_SEC = 15;
 
@@ -20,29 +20,230 @@ const IGNORED_TASK_DEFINITION_ATTRIBUTES = [
   'registeredBy'
 ];
 
+// Method to run a stand-alone task with desired inputs
+async function runTask(ecs, clusterName, taskDefArn, waitForMinutes, enableECSManagedTags) {
+  core.info('Running task')
+
+  const waitForTask = core.getInput('wait-for-task-stopped', { required: false }) || 'false';
+  const startedBy = core.getInput('run-task-started-by', { required: false }) || 'GitHub-Actions';
+  const launchType = core.getInput('run-task-launch-type', { required: false }) || 'FARGATE';
+  const subnetIds = core.getInput('run-task-subnets', { required: false }) || '';
+  const securityGroupIds = core.getInput('run-task-security-groups', { required: false }) || '';
+  const containerOverrides = JSON.parse(core.getInput('run-task-container-overrides', { required: false }) || '[]');
+  const assignPublicIP = core.getInput('run-task-assign-public-IP', { required: false }) || 'DISABLED';
+  const tags = JSON.parse(core.getInput('run-task-tags', { required: false }) || '[]');
+  const capacityProviderStrategy = JSON.parse(core.getInput('run-task-capacity-provider-strategy', { required: false }) || '[]');
+  const runTaskManagedEBSVolumeName = core.getInput('run-task-managed-ebs-volume-name', { required: false }) || '';
+  const runTaskManagedEBSVolume = core.getInput('run-task-managed-ebs-volume', { required: false }) || '{}';
+
+  let awsvpcConfiguration = {}
+
+  if (subnetIds != "") {
+    awsvpcConfiguration["subnets"] = subnetIds.split(',')
+  }
+
+  if (securityGroupIds != "") {
+    awsvpcConfiguration["securityGroups"] = securityGroupIds.split(',')
+  }
+
+  if(assignPublicIP != "" && (subnetIds != "" || securityGroupIds != "")){
+    awsvpcConfiguration["assignPublicIp"] = assignPublicIP
+  }
+  let volumeConfigurations = [];
+  let taskManagedEBSVolumeObject;
+
+  if (runTaskManagedEBSVolumeName != '') {
+    if (runTaskManagedEBSVolume != '{}') {
+      taskManagedEBSVolumeObject = convertToManagedEbsVolumeObject(runTaskManagedEBSVolume);
+      volumeConfigurations = [{
+            name: runTaskManagedEBSVolumeName,
+            managedEBSVolume: taskManagedEBSVolumeObject
+      }];
+    } else {
+      core.warning(`run-task-managed-ebs-volume-name provided without run-task-managed-ebs-volume value. VolumeConfigurations property will not be included in the RunTask API call`);
+    }
+  }
+
+  const runTaskResponse = await ecs.runTask({
+    startedBy: startedBy,
+    cluster: clusterName,
+    taskDefinition: taskDefArn,
+    overrides: {
+      containerOverrides: containerOverrides
+    },
+    capacityProviderStrategy: capacityProviderStrategy.length === 0 ? null : capacityProviderStrategy,
+    launchType: capacityProviderStrategy.length === 0 ? launchType : null,
+    networkConfiguration: Object.keys(awsvpcConfiguration).length === 0 ? null : { awsvpcConfiguration: awsvpcConfiguration },
+    enableECSManagedTags: enableECSManagedTags,
+    tags: tags,
+    volumeConfigurations: volumeConfigurations
+  });
+
+  core.debug(`Run task response ${JSON.stringify(runTaskResponse)}`)
+
+  const taskArns = runTaskResponse.tasks.map(task => task.taskArn);
+  core.setOutput('run-task-arn', taskArns);
+
+  const region = await ecs.config.region();
+  const consoleHostname = region.startsWith('cn') ? 'console.amazonaws.cn' : 'console.aws.amazon.com';
+
+  core.info(`Task running: https://${consoleHostname}/ecs/home?region=${region}#/clusters/${clusterName}/tasks`);
+
+  if (runTaskResponse.failures && runTaskResponse.failures.length > 0) {
+    const failure = runTaskResponse.failures[0];
+    throw new Error(`${failure.arn} is ${failure.reason}`);
+  }
+
+  // Wait for task to end
+  if (waitForTask && waitForTask.toLowerCase() === "true") {
+    await waitForTasksStopped(ecs, clusterName, taskArns, waitForMinutes)
+    await tasksExitCode(ecs, clusterName, taskArns)
+  } else {
+    core.debug('Not waiting for the task to stop');
+  }
+}
+
+function convertToManagedEbsVolumeObject(managedEbsVolume) {
+  managedEbsVolumeObject = {}
+  const ebsVolumeObject = JSON.parse(managedEbsVolume);
+  if ('roleArn' in ebsVolumeObject){ // required property
+    managedEbsVolumeObject.roleArn = ebsVolumeObject.roleArn;
+    core.debug(`Found RoleArn ${ebsVolumeObject['roleArn']}`);
+  } else {
+    throw new Error('managed-ebs-volume must provide "role-arn" to associate with the EBS volume')
+  }
+
+  if ('encrypted' in ebsVolumeObject) {
+    managedEbsVolumeObject.encrypted = ebsVolumeObject.encrypted;
+  }
+  if ('filesystemType' in ebsVolumeObject) {
+    managedEbsVolumeObject.filesystemType = ebsVolumeObject.filesystemType;
+  }
+  if ('iops' in ebsVolumeObject) {
+    managedEbsVolumeObject.iops = ebsVolumeObject.iops;
+  }
+  if ('kmsKeyId' in ebsVolumeObject) {
+    managedEbsVolumeObject.kmsKeyId = ebsVolumeObject.kmsKeyId;
+  }
+  if ('sizeInGiB' in ebsVolumeObject) {
+    managedEbsVolumeObject.sizeInGiB = ebsVolumeObject.sizeInGiB;
+  }
+  if ('snapshotId' in ebsVolumeObject) {
+    managedEbsVolumeObject.snapshotId = ebsVolumeObject.snapshotId;
+  }
+  if ('tagSpecifications' in ebsVolumeObject) {
+    managedEbsVolumeObject.tagSpecifications = ebsVolumeObject.tagSpecifications;
+  }
+  if (('throughput' in ebsVolumeObject) && (('volumeType' in ebsVolumeObject) && (ebsVolumeObject.volumeType == 'gp3'))){
+    managedEbsVolumeObject.throughput = ebsVolumeObject.throughput;
+  }
+  if ('volumeType' in ebsVolumeObject) {
+    managedEbsVolumeObject.volumeType = ebsVolumeObject.volumeType;
+  }
+  core.debug(`Created managedEbsVolumeObject: ${JSON.stringify(managedEbsVolumeObject)}`);
+  return managedEbsVolumeObject;
+}
+
+// Poll tasks until they enter a stopped state
+async function waitForTasksStopped(ecs, clusterName, taskArns, waitForMinutes) {
+  if (waitForMinutes > MAX_WAIT_MINUTES) {
+    waitForMinutes = MAX_WAIT_MINUTES;
+  }
+
+  core.info(`Waiting for tasks to stop. Will wait for ${waitForMinutes} minutes`);
+
+  const waitTaskResponse = await waitUntilTasksStopped({
+    client: ecs,
+    minDelay: WAIT_DEFAULT_DELAY_SEC,
+    maxWaitTime: waitForMinutes * 60,
+  }, {
+    cluster: clusterName,
+    tasks: taskArns,
+  });
+
+  core.debug(`Run task response ${JSON.stringify(waitTaskResponse)}`);
+  core.info('All tasks have stopped.');
+}
+
+// Check a task's exit code and fail the job on error
+async function tasksExitCode(ecs, clusterName, taskArns) {
+  const describeResponse = await ecs.describeTasks({
+    cluster: clusterName,
+    tasks: taskArns
+  });
+
+  const containers = [].concat(...describeResponse.tasks.map(task => task.containers))
+  const exitCodes = containers.map(container => container.exitCode)
+  const reasons = containers.map(container => container.reason)
+
+  const failuresIdx = [];
+
+  exitCodes.filter((exitCode, index) => {
+    if (exitCode !== 0) {
+      failuresIdx.push(index)
+    }
+  })
+
+  const failures = reasons.filter((_, index) => failuresIdx.indexOf(index) !== -1)
+  if (failures.length > 0) {
+    throw new Error(`Run task failed: ${JSON.stringify(failures)}`);
+  }
+}
+
 // Deploy to a service that uses the 'ECS' deployment controller
-async function updateEcsService(ecs, clusterName, service, taskDefArn, waitForService, waitForMinutes, forceNewDeployment) {
+async function updateEcsService(ecs, clusterName, service, taskDefArn, waitForService, waitForMinutes, forceNewDeployment, desiredCount, enableECSManagedTags, propagateTags) {
   core.debug('Updating the service');
-  await ecs.updateService({
+
+  const serviceManagedEBSVolumeName = core.getInput('service-managed-ebs-volume-name', { required: false }) || '';
+  const serviceManagedEBSVolume = core.getInput('service-managed-ebs-volume', { required: false }) || '{}';
+
+  let volumeConfigurations = [];
+  let serviceManagedEbsVolumeObject;
+
+  if (serviceManagedEBSVolumeName != '') {
+    if (serviceManagedEBSVolume != '{}') {
+      serviceManagedEbsVolumeObject = convertToManagedEbsVolumeObject(serviceManagedEBSVolume);
+      volumeConfigurations = [{
+        name: serviceManagedEBSVolumeName,
+        managedEBSVolume: serviceManagedEbsVolumeObject
+      }];
+    } else {
+      core.warning('service-managed-ebs-volume-name provided without service-managed-ebs-volume value. VolumeConfigurations property will not be included in the UpdateService API call');
+    }
+  }
+
+  let params = {
     cluster: clusterName,
     service: service,
     taskDefinition: taskDefArn,
-    forceNewDeployment: forceNewDeployment
-  }).promise();
-  core.info(`Deployment started. Watch this deployment's progress in the Amazon ECS console: https://console.aws.amazon.com/ecs/home?region=${aws.config.region}#/clusters/${clusterName}/services/${service}/events`);
+    forceNewDeployment: forceNewDeployment,
+    enableECSManagedTags: enableECSManagedTags,
+    propagateTags: propagateTags,
+    volumeConfigurations: volumeConfigurations
+  };
+
+  // Add the desiredCount property only if it is defined and a number.
+  if (!isNaN(desiredCount) && desiredCount !== undefined) {
+    params.desiredCount = desiredCount;
+  }
+  await ecs.updateService(params);
+
+  const region = await ecs.config.region();
+  const consoleHostname = region.startsWith('cn') ? 'console.amazonaws.cn' : 'console.aws.amazon.com';
+
+  core.info(`Deployment started. Watch this deployment's progress in the Amazon ECS console: https://${region}.${consoleHostname}/ecs/v2/clusters/${clusterName}/services/${service}/events?region=${region}`);
 
   // Wait for service stability
   if (waitForService && waitForService.toLowerCase() === 'true') {
     core.debug(`Waiting for the service to become stable. Will wait for ${waitForMinutes} minutes`);
-    const maxAttempts = (waitForMinutes * 60) / WAIT_DEFAULT_DELAY_SEC;
-    await ecs.waitFor('servicesStable', {
+    await waitUntilServicesStable({
+      client: ecs,
+      minDelay: WAIT_DEFAULT_DELAY_SEC,
+      maxWaitTime: waitForMinutes * 60
+    }, {
       services: [service],
-      cluster: clusterName,
-      $waiter: {
-        delay: WAIT_DEFAULT_DELAY_SEC,
-        maxAttempts: maxAttempts
-      }
-    }).promise();
+      cluster: clusterName
+    });
   } else {
     core.debug('Not waiting for the service to become stable');
   }
@@ -174,10 +375,12 @@ async function createCodeDeployDeployment(codedeploy, clusterName, service, task
 
   let codeDeployDescription = core.getInput('codedeploy-deployment-description', { required: false });
 
+  let codeDeployConfig = core.getInput('codedeploy-deployment-config', { required: false });
+
   let deploymentGroupDetails = await codedeploy.getDeploymentGroup({
     applicationName: codeDeployApp,
     deploymentGroupName: codeDeployGroup
-  }).promise();
+  });
   deploymentGroupDetails = deploymentGroupDetails.deploymentGroupInfo;
 
   // Insert the task def ARN into the appspec file
@@ -212,13 +415,20 @@ async function createCodeDeployDeployment(codedeploy, clusterName, service, task
       }
     }
   };
+
   // If it hasn't been set then we don't even want to pass it to the api call to maintain previous behaviour.
   if (codeDeployDescription) {
-    deploymentParams.description = codeDeployDescription
+    // CodeDeploy Deployment Descriptions have a max length of 512 characters, so truncate if necessary
+    deploymentParams.description = (codeDeployDescription.length <= 512) ? codeDeployDescription : `${codeDeployDescription.substring(0,511)}…`;
   }
-  const createDeployResponse = await codedeploy.createDeployment(deploymentParams).promise();
+  if (codeDeployConfig) {
+    deploymentParams.deploymentConfigName = codeDeployConfig
+  }
+  const createDeployResponse = await codedeploy.createDeployment(deploymentParams);
   core.setOutput('codedeploy-deployment-id', createDeployResponse.deploymentId);
-  core.info(`Deployment started. Watch this deployment's progress in the AWS CodeDeploy console: https://console.aws.amazon.com/codesuite/codedeploy/deployments/${createDeployResponse.deploymentId}?region=${aws.config.region}`);
+
+  const region = await codedeploy.config.region();
+  core.info(`Deployment started. Watch this deployment's progress in the AWS CodeDeploy console: https://console.aws.amazon.com/codesuite/codedeploy/deployments/${createDeployResponse.deploymentId}?region=${region}`);
 
   // Wait for deployment to complete
   if (waitForService && waitForService.toLowerCase() === 'true') {
@@ -229,16 +439,14 @@ async function createCodeDeployDeployment(codedeploy, clusterName, service, task
     if (totalWaitMin > MAX_WAIT_MINUTES) {
       totalWaitMin = MAX_WAIT_MINUTES;
     }
-    const maxAttempts = (totalWaitMin * 60) / WAIT_DEFAULT_DELAY_SEC;
-
     core.debug(`Waiting for the deployment to complete. Will wait for ${totalWaitMin} minutes`);
-    await codedeploy.waitFor('deploymentSuccessful', {
-      deploymentId: createDeployResponse.deploymentId,
-      $waiter: {
-        delay: WAIT_DEFAULT_DELAY_SEC,
-        maxAttempts: maxAttempts
-      }
-    }).promise();
+    await waitUntilDeploymentSuccessful({
+      client: codedeploy,
+      minDelay: WAIT_DEFAULT_DELAY_SEC,
+      maxWaitTime: totalWaitMin * 60
+    }, {
+      deploymentId: createDeployResponse.deploymentId
+    });
   } else {
     core.debug('Not waiting for the deployment to complete');
   }
@@ -246,10 +454,10 @@ async function createCodeDeployDeployment(codedeploy, clusterName, service, task
 
 async function run() {
   try {
-    const ecs = new aws.ECS({
+    const ecs = new ECS({
       customUserAgent: 'amazon-ecs-deploy-task-definition-for-github-actions'
     });
-    const codedeploy = new aws.CodeDeploy({
+    const codedeploy = new CodeDeploy({
       customUserAgent: 'amazon-ecs-deploy-task-definition-for-github-actions'
     });
 
@@ -259,12 +467,20 @@ async function run() {
     const cluster = core.getInput('cluster', { required: false });
     const waitForService = core.getInput('wait-for-service-stability', { required: false });
     let waitForMinutes = parseInt(core.getInput('wait-for-minutes', { required: false })) || 30;
+
     if (waitForMinutes > MAX_WAIT_MINUTES) {
       waitForMinutes = MAX_WAIT_MINUTES;
     }
 
     const forceNewDeployInput = core.getInput('force-new-deployment', { required: false }) || 'false';
     const forceNewDeployment = forceNewDeployInput.toLowerCase() === 'true';
+    const desiredCount = parseInt((core.getInput('desired-count', {required: false})));
+    const enableECSManagedTagsInput = core.getInput('enable-ecs-managed-tags', { required: false }) || '';
+    let enableECSManagedTags = null;
+    if (enableECSManagedTagsInput !== '') {
+      enableECSManagedTags = enableECSManagedTagsInput.toLowerCase() === 'true';
+    }
+    const propagateTags = core.getInput('propagate-tags', { required: false }) || 'NONE';
 
     // Register the task definition
     core.debug('Registering the task definition');
@@ -275,7 +491,7 @@ async function run() {
     const taskDefContents = maintainValidObjects(removeIgnoredAttributes(cleanNullKeys(yaml.parse(fileContents))));
     let registerResponse;
     try {
-      registerResponse = await ecs.registerTaskDefinition(taskDefContents).promise();
+      registerResponse = await ecs.registerTaskDefinition(taskDefContents);
     } catch (error) {
       core.setFailed("Failed to register task definition in ECS: " + error.message);
       core.debug("Task definition contents:");
@@ -285,15 +501,23 @@ async function run() {
     const taskDefArn = registerResponse.taskDefinition.taskDefinitionArn;
     core.setOutput('task-definition-arn', taskDefArn);
 
+    // Run the task outside of the service
+    const clusterName = cluster ? cluster : 'default';
+    const shouldRunTaskInput = core.getInput('run-task', { required: false }) || 'false';
+    const shouldRunTask = shouldRunTaskInput.toLowerCase() === 'true';
+    core.debug(`shouldRunTask: ${shouldRunTask}`);
+    if (shouldRunTask) {
+      core.debug("Running ad-hoc task...");
+      await runTask(ecs, clusterName, taskDefArn, waitForMinutes, enableECSManagedTags);
+    }
+
     // Update the service with the new task definition
     if (service) {
-      const clusterName = cluster ? cluster : 'default';
-
       // Determine the deployment controller
       const describeResponse = await ecs.describeServices({
         services: [service],
         cluster: clusterName
-      }).promise();
+      });
 
       if (describeResponse.failures && describeResponse.failures.length > 0) {
         const failure = describeResponse.failures[0];
@@ -305,11 +529,14 @@ async function run() {
         throw new Error(`Service is ${serviceResponse.status}`);
       }
 
-      if (!serviceResponse.deploymentController) {
+      if (!serviceResponse.deploymentController || !serviceResponse.deploymentController.type || serviceResponse.deploymentController.type === 'ECS') {
         // Service uses the 'ECS' deployment controller, so we can call UpdateService
-        await updateEcsService(ecs, clusterName, service, taskDefArn, waitForService, waitForMinutes, forceNewDeployment);
-      } else if (serviceResponse.deploymentController.type == 'CODE_DEPLOY') {
+        core.debug('Updating service...');
+        await updateEcsService(ecs, clusterName, service, taskDefArn, waitForService, waitForMinutes, forceNewDeployment, desiredCount, enableECSManagedTags, propagateTags);
+
+      } else if (serviceResponse.deploymentController.type === 'CODE_DEPLOY') {
         // Service uses CodeDeploy, so we should start a CodeDeploy deployment
+        core.debug('Deploying service in the default cluster');
         await createCodeDeployDeployment(codedeploy, clusterName, service, taskDefArn, waitForService, waitForMinutes);
       } else {
         throw new Error(`Unsupported deployment controller: ${serviceResponse.deploymentController.type}`);